backend/map-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.30
aiomysql==0.2.0
pydantic-settings==2.5.0
